Remove atmos.py module
