+++ pyproject.toml
@@ -40,7 +40,7 @@ Changelog = "https://github.com/actris-cloudnet/mwrpy/blob/main/CHANGELOG.md"
 check_untyped_defs = true
 
 [[tool.mypy.overrides]]
-module = ["cftime.*", "mpl_toolkits.*", "pandas.*", "scipy.*", "suncalc.*"]
+module = ["atmoslib.*", "matplotlib.*", "netCDF4.*", "pandas.*", "scipy.*", "suncalc.*", "timezonefinder.*"]
 ignore_missing_imports = true
 
 [tool.release-version]
